ProjectForWork/work/CsvReader/requirements.txt
pandas
python-docx
python-dateutil
